+++ lib/tawkto_flutter.dart
@@ -1,7 +1,4 @@
 library tawkto_flutter;
 
-/// A Calculator.
-class Calculator {
-  /// Returns [value] plus 1.
-  int addOne(int value) => value + 1;
-}
+export 'src/tawkto_visitor.dart';
+export 'src/tawkto_widget.dart';